# docs/swagger.yaml
basePath: /x_ray
definitions:
  dto.CompanyRes:
    properties:
      album:
        items:
          type: string
        type: array
      created_at:
        type: string
      description:
        type: string
      headquarters:
        type: string
      id:
        type: string
      industry:
        type: string
      logo:
        type: string
      name:
        type: string
      size:
        type: string
      updated_at:
        type: string
      website:
        type: string
    type: object
...
  dto.UserRes:
    properties:
      avatar_url:
        type: string
      created_at:
        type: integer
      email:
        type: string
      google_id:
        type: string
      id:
        type: string
      last_login:
        type: integer
      name:
        type: string
      position:
        type: string
      updated_at:
        type: integer
    type: object
  localization.StandardResponse:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
      time:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: this Backend is for x-ray, a company review platform. this backend/project
    was build with the intention to providing transparency to the job seekers, and
    also to give the company a chance to improve their work environment. this project
    is still in early stage, so expect some breaking changes in the future.
  title: X-ray API
  version: "1.0"
paths:
  /auth/callback:
    get:
      consumes:
      - application/json
      description: Handles the OAuth callback from Google, exchanges authorization
        code for tokens, and creates/authenticates user
      parameters:
      - description: Authorization code from Google
        in: query
        name: code
        required: true
        type: string
      - description: State parameter for CSRF protection
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful login with user data
          schema:
            allOf:
            - $ref: '#/definitions/localization.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserRes'
              type: object
        "400":
          description: Invalid callback parameters or state mismatch
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "500":
          description: Internal server error during OAuth process
          schema:
            $ref: '#/definitions/localization.StandardResponse'
      summary: Google OAuth Callback
      tags:
      - Auth
  /auth/login:
    get:
      description: Redirects user to Google OAuth consent screen
      produces:
      - application/json
      responses:
        "302":
          description: Redirect to Google OAuth
          schema:
            type: string
        "500":
          description: Missing environment variables
          schema:
            $ref: '#/definitions/localization.StandardResponse'
      summary: Initiate Google OAuth Login
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Generates a new access token from a valid refresh token
      produces:
      - application/json
      responses:
        "200":
          description: New access token generated
          schema:
            allOf:
            - $ref: '#/definitions/localization.StandardResponse'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized - invalid or missing refresh token
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/localization.StandardResponse'
      security:
      - BearerAuth: []
      summary: Refresh Access Token
      tags:
      - Auth
  /company:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of companies with optional filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Companies fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/localization.StandardResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.CompanyRes'
                  type: array
              type: object
        "400":
          description: Invalid filter parameters
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/localization.StandardResponse'
      summary: Get Companies by Pagination
      tags:
      - Company
    post:
      consumes:
      - multipart/form-data
      description: Creates a new company with details, logo, and album images (multipart
        form data)
      parameters:
      - description: Company name
        in: formData
        name: name
        required: true
        type: string
      - description: Company website URL
        in: formData
        name: website
        type: string
      - description: Industry type
        in: formData
        name: industry
        type: string
      - description: Company size
        in: formData
        name: size
        type: string
      - description: Headquarters location
        in: formData
        name: headquarters
        type: string
      - description: Company description
        in: formData
        name: description
        type: string
      - description: Company logo image
        in: formData
        name: logo
        type: file
      - description: Company album images (multiple files)
        in: formData
        name: album
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Company created successfully
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "400":
          description: Invalid request body or form data
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "401":
          description: Unauthorized - user not authenticated
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/localization.StandardResponse'
      security:
      - BearerAuth: []
      summary: Create Company
      tags:
      - Company
  /company/detail:
    get:
      consumes:
      - application/json
      description: Retrieves detailed information about a specific company
      parameters:
      - description: Company ID (UUID)
        in: query
        name: company_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Company fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/localization.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.CompanyRes'
              type: object
        "400":
          description: Missing company_id parameter
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "404":
          description: Company not found
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/localization.StandardResponse'
      summary: Get Company by ID
      tags:
      - Company
...
swagger: "2.0"